Changed the approach a bit.
I found that the previous approach did not behave well when moving from having
an environmental variable to not having one. While this would be ok
on a transient instance, it would break a local install.

Instead, this uses a temporary system property to get the data source setup,
then the property is always passed in to standalone when the application is run.

This seems to work well locally.